控制编译大小的宏

+++ include/kernel/interrupt/idt.h
@@ -9,7 +9,7 @@
 #define IDT_DESC_ATTR(p, dpl, type) ((p << 7) | (dpl << 5) | (type))
 
 #define IDT_GET_PTR(idt_addr)                                                  \
-  (((u64)(u32)idt_addr << 16) | (sizeof(idt_addr) - 1))
+  (((u64)(uintptr_t)idt_addr << 16) | (sizeof(idt_addr) - 1))
 
 /**
  * @brief Initializes the IDT
